+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_errored_definition.php
@@ -494,6 +494,11 @@ class Symfony_DI_PhpDumper_Errored_Definition extends Container
         throw new ParameterNotFoundException($name);
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [