+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_rot13_env.php
@@ -117,6 +117,11 @@ class Symfony_DI_PhpDumper_Test_Rot13Parameters extends Container
         return $this->dynamicParameters[$name] = $value;
     }
 
+    public function isParameterDynamic(string $name): bool
+    {
+        return isset($this->loadedDynamicParameters[$name]);
+    }
+
     protected function getDefaultParameters(): array
     {
         return [